tests/wpt/tests/html/canvas/element/layers/2d.layer.exceptions-are-no-op.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.layer.exceptions-are-no-op</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.layer.exceptions-are-no-op</h1>
<p class="desc">Checks that the context state is left unchanged if beginLayer throws.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Checks that the context state is left unchanged if beginLayer throws.");
_addTest(function(canvas, ctx) {

  // Get `beginLayer` to throw while parsing the filter.
  assert_throws_js(TypeError,
                   () => ctx.beginLayer({filter: {name: 'colorMatrix',
                                                  values: 'foo'}}));
  // `beginLayer` shouldn't have opened the layer, so `endLayer` should throw.
  assert_throws_dom("InvalidStateError", () => ctx.endLayer());

});
</script>
